Protect against empty APNS certificate
By default (as defined in the env.development file), P8_CERTIFICATE is an empty string. This causes a fatal error when attempting to setup the APNS configuration.

Since APNS is not an essential component, at least in the development environment, I've added some protection here to make it easier for developers to get started.