Disable Postgres TLS override
I was receiving a number of "SSL Unsupported" errors preventing a successful startup. This change fixes it.

It does appear on the face of it that this is simply disabling TLS/SSL, this is not entirely the case as (as per the documentation on the PostgresConfiguration initialiser) that if the URL provided has either "ssl=true" or "sslmode=require" within the query parameters then the functionality is identical.

As such, when using a correctly construced URL in production, this should have no impact on clients and prevent errors for the rest of us.