Add Seed Capability
Currently the admin section is undocumented and not easy to access. By seeding an admin user with documented values, I hope this will lower the hurdle for many.

We may wish to add more commands to the seed SQL document to fill out the schedule - perhaps based on the data from 2022.