Remove unused image field on presentation
I think there is value is an image attachd to a presentation, but I think this makes more sense as an array when we add galleries in the future to showcase previous talks.

This field is seemingly never populated, and is not used by either the app or the website (which uses the speaker image instead).