Add SwiftFormat, and run it
A little opinionated, but has definitely cleaned up a good amount of the code (just silly indentation and such).

Not enforcing it at the moment, but have left the configuration in for anyone who wants to format with the same rules.

Maybe want to add a header? Most OSS packages add the license there.