Add AASA unit test
I feel priviledged to write the first unit test 😂 It's a bit of a niche case as this route is defined purely in middleware and as such is a bit hidden. We may want to consider moving this to an actual route - but if we do, this test should continue to work without changes.