allow for admin control over slot dates
this now uses a proper date selector, removing the time input before.

the time input will still be populated for now (for backwards compatibility) but can be removed in the future.

if you edit and save, without changing, an existing slot - it will simply populate the correct time in the database based on the date of the first SwiftLeeds.