Add year switcher beta, always show sponsors
All context builders now check for the injected event rather than for 'isCurrent'. This allows us to specify other events as part of the previous year navigator.

I've added a temporary year chooser in the form of a ?year=2023 query parameter. This is not finished, and thus is only on debug builds.

I've also removed the sponsors phase management as we would never want to hide this. The call to action to become a sponsor is always shown, and the sponsors grid itself is based on the existence of any sponsor in the db.