Add support for Reusable Activities
A reusable activity is an activity without an event associated with it.

They're useful for repeatable activities such as lunch, or registration where the details do not change year to year.

Reusable activities will show regardless of your year filter, as they can be used with any year.

This also changes the slot admin form to hide non-supported activities and events rather than simply making them disabled. This is because the list is getting rather long with all the presentations from every year.