Add Day Association to Slots
I wanted to remove EventID but this turned into a horrible sequence of operations to make it work.

As it stands, we now populate both the DayID and the EventID. This will change slightly in this PR, but will also be cleared up fully in a future PR once we're not afraid of production.