Recycle v1 name after refactor
All the old v1 models/transformers/etc were not used in the codebase, so could be deleted. This freed the namespace up for the 'v2' ones to be called 'v1' which aligns with the API version (api/v1/schedule etc).

This is helpful as we're now making v2 of this API, which would be confusing to use v3 of the models.